Fix buffer overflow bug in deflate::core::BitBuffer::flush
The problem is that this function does not check that there is enough space to fit a u64, it only checks that the starting position for the write is in bounds.
I have tried adding `assert!(pos.checked_add(8).unwrap() <= output.inner.get_ref().len());` before the unsafe part, but that regressed performance exactly as much as the fully safe version, so I went for the fully safe code.