Flag `miniz_oxide` as a `#![no_std]` library (#81)
* Flag `miniz_oxide` as a `#![no_std]` library

This commit conversion the `miniz_oxide` crate to a `#![no_std]`
library, and is as a result a breaking change for the library. Currently
the only dependency on the `std` crate is the `std::io::Cursor` type,
which is pretty easily replaced with a `usize` parameter in a few
locations.

The goal of this commit is to eventually enable this library to be
included into the standard library itself. Dependencies of the standard
library can't depend on the standard library as well! The reason for
including this in the standard library is that the `backtrace` crate
wants to decompress DWARF information in executables, which can be
compressed with zlib.

* Update with annotations to build in libstd