Derive Clone for InflateState to allow random-access reads (#157)
I am implementing Seek for a wrapper that reads deflated files.  To do this, I will
inflate the file once and save the dictionary and inflate state at several points in
the file.  Later, when I want to seek to some random place in the file, I will use the
cloned state+dict to position to a known file position and then inflate until I reach
the data I need.

To support this, I need to be able to Clone the InflateState and its parts (the current
dictionary).  Derive Clone for InflateState, HuffmanTable and DecompressorOxide.